Commented last $arrayRoute parameter check
In our case, this last check is redundant and makes our app work in a wrong way. Because we don't want to have our app behave that way.

You can think about this check as a strict one. Because it checks everything till the end. In our case we need to skip last part. It will work like a category for the rest of a pages.